+++ figures-counter/includes/error.h
@@ -18,41 +18,41 @@ public:
 
 class row_length_error : public bad_input {
 public:
-	row_length_error(unsigned row, unsigned length, unsigned expected_length);
+	row_length_error(size_t row, size_t length, size_t expected_length);
 
-	unsigned row() const noexcept;
-	unsigned length() const noexcept;
-	unsigned expected_length() const noexcept;
+	size_t row() const noexcept;
+	size_t length() const noexcept;
+	size_t expected_length() const noexcept;
 
 private:
-	unsigned _row;
-	unsigned _length;
-	unsigned _expected_length;
+	size_t _row;
+	size_t _length;
+	size_t _expected_length;
 };
 
 class row_count_error : public bad_input {
 public:
-	row_count_error(unsigned count, unsigned expected_count);
+	row_count_error(size_t count, size_t expected_count);
 
-	unsigned count() const noexcept;
-	unsigned expected_count() const noexcept;
+	size_t count() const noexcept;
+	size_t expected_count() const noexcept;
 
 private:
-	unsigned _count;
-	unsigned _expected_count;
+	size_t _count;
+	size_t _expected_count;
 };
 
 class unrecognized_symbol : public bad_input {
 public:
-	unrecognized_symbol(unsigned row, unsigned col, char symbol);
+	unrecognized_symbol(size_t row, size_t col, char symbol);
 
-	unsigned row() const noexcept;
-	unsigned col() const noexcept;
+	size_t row() const noexcept;
+	size_t col() const noexcept;
 	char symbol() const noexcept;
 
 private:
-	unsigned _row;
-	unsigned _col;
+	size_t _row;
+	size_t _col;
 	char _symbol;
 };
